// shared/src/interfaces/applicationInterface.ts
/**
 * @fileoverview Defines the IApplication interface for application data.
 * @version 1.0.0
 * @module applicationInterface
 */

import { Optional } from 'sequelize';

/**
 * Interface representing the attributes of an application.
 */
export interface IApplication {
  id?: number;
  jobId: number;
  candidateId: number;
  status?: 'Pending' | 'Approved' | 'Shortlisted' | 'Rejected';
  createdAt?: Date;
  updatedAt?: Date;
}

/**
 * Interface for creating a new application (excluding autogenerated fields).
 */
export type IApplicationCreation = Optional<
  IApplication,
  'id' | 'status' | 'createdAt' | 'updatedAt'
>;